pyproject.toml
[tool.poetry]
name = "bookstore"
version = "2.2"
description = ""
authors = ["Juliana Rodrigues <juliana.rodrigues-29@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.4"
djangorestframework = "^3.15.2"
django-extensions = "^3.2.3"
coreapi = "^2.3.3"
django-debug-toolbar = "^4.4.6"
whitenoise = "^6.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
